News Archive
This is an experiment to see if (1) the site admins like it and (2) I executed it correctly. I attempted to create a new page (News Archive) to house older news posts--2016 thru Oct 2018. Once 2019 news accumulates for a few months, I'll move the rest of 2018 news into the archive. I figured I would ask forgiveness instead of permission.

1. Created "news" folder under "llnl.github.io" and moved news.md into it
2. Created "newsarchive" folder under "_posts" & moved older posts into it
3. Created news-archive.md (in new "news" folder) using default layout for news
4. Added link on news.md to News Archive
5. Updated header.html to turn News into a dropdown menu with News Archive as subordinate option

The thing I am least sure about is whether line 7 in news-archive.md will correctly pull in only the posts in the "newsarchive" subfolder: {% for page in site.posts.newsarchive %}